requirements.txt
flask
flask-cors
requests
python-dotenv
jinja2
pyyaml